Completed first CSS methods exercise
This lesson focused on adding styles to a html document via the
three main methods: inline, internal , and external methods. The
inline method is done by actually adding a style attribute to an
element's tag. While the internal and external methods both have
use sectors to specify which elements to modify. The internal method
style settings are located in the head of the html file. In contrast,
the external method applies styles to a html file via a seperate css
file and a link element with attributes the role and path of the file.