Complete CSS lessons on margin and padding
These lessons focused on manipulating a website using CSS declarations
of margin and padding. Margin and padding declarations can be applied
to vertical and horizontal with out seperate calls i.e margin-top and
margin-right. example:

margin : 6px 8px ;

will apply a margin of 6px to top and bottom and 8px to right and left.
The same can be done with padding.

Also used in the lesson was the display declaration to change an inline
element to a block element.